ci: verify presence of valid authenticode signature
In the current verification, code checks for only presence of the signature which can be problematic if the signature is no longer valid.

`Get-AuthenticodeSignature` allows to check whether the signature is valid or not. This is also covered at https://devblogs.microsoft.com/scripting/reporting-on-digitally-signed-files-with-powershell/

This change replaces the file read with signature validity check using `Get-AuthenticodeSignature` cmdlet.

Invalid signature module files now fail the validity pass.